+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to
 * Replaced `EthernetClient::writeFully(const char *, size_t)` and
   `writeFully(const uint8_t *, size_t)` with `writeFully(const void *, size_t)`.
 * Updated the file structure to put most sources underneath _src/qnethernet/_.
+* Changed `EthernetClient::setConnectionTimeout(timeout)` and
+  `connectionTimeout()` to use 32-bit values.
 
 ### Fixed
 * Now using `(void)` instead of `()` for all C function declarations because